refactor: LottoMachine이 구입 금액에 대한 유효성 검사하는 것을 Money 객체가 하도록 생성 및 위임
